+++ malachite-float/src/test_util/generators/special_random.rs
@@ -27,7 +27,8 @@ use crate::test_util::generators::exhaustive::{
     acos_prec_round_valid, acos_rational_prec_round_valid, acos_round_valid,
     acos_with_period_prec_round_valid, acos_with_period_rational_prec_round_valid,
     acos_with_period_round_valid, acsc_prec_round_valid, acsc_rational_prec_round_valid,
-    acsc_round_valid, acsc_with_period_prec_round_valid, acsc_with_period_round_valid,
+    acsc_round_valid, acsc_with_period_prec_round_valid,
+    acsc_with_period_rational_prec_round_valid, acsc_with_period_round_valid,
     add_mul_prec_round_valid, add_mul_rational_prec_round_valid, add_mul_rational_round_valid,
     add_mul_round_valid, add_prec_round_valid, add_rational_prec_round_valid,
     add_rational_round_valid, add_round_valid, agm_prec_round_valid, agm_rational_prec_round_valid,
@@ -14179,5 +14180,40 @@ pub fn special_random_float_unsigned_rounding_mode_triple_gen_var_50(
             &random_rounding_modes,
         )
         .filter(|&(ref x, u, rm)| acsc_with_period_round_valid(x, u, rm)),
+    )
+}
+
+pub fn special_random_rational_unsigned_unsigned_rounding_mode_quadruple_gen_var_10(
+    config: &GenConfig,
+) -> It<(Rational, u64, u64, RoundingMode)> {
+    Box::new(
+        random_quadruples(
+            EXAMPLE_SEED,
+            &|seed| {
+                striped_random_rationals(
+                    seed,
+                    config.get_or("mean_stripe_n", 32),
+                    config.get_or("mean_stripe_d", 1),
+                    config.get_or("mean_bits_n", 64),
+                    config.get_or("mean_bits_d", 1),
+                )
+            },
+            &|seed| {
+                geometric_random_unsigneds(
+                    seed,
+                    config.get_or("mean_base_n", 16),
+                    config.get_or("mean_base_d", 1),
+                )
+            },
+            &|seed| {
+                geometric_random_positive_unsigneds(
+                    seed,
+                    config.get_or("small_unsigned_mean_n", 4),
+                    config.get_or("small_unsigned_mean_d", 1),
+                )
+            },
+            &random_rounding_modes,
+        )
+        .filter(|&(ref n, k, prec, rm)| acsc_with_period_rational_prec_round_valid(n, k, prec, rm)),
     )
 }